==> notebooks/1.0-EDA.ipynb <==
import pandas as pd
import string

Firsly, take a look at the data:

df = pd.read_csv('../data/raw/filtered.tsv', delimiter='\t', index_col=0)
df

# Basic Preprocessing

We will remove punctuation signs, however the meaningful apostrophe character should remain.

def basic_preprocessing(text: str) -> str:
    # remove punctuation
    text = text.translate(str.maketrans('', '', string.punctuation.replace("'", "")))
    # lowercase
    text = text.lower()
    return text

# apply basic preprocessing
for column in ['reference', 'translation']:
    df[f'{column}_processed'] = df[column].apply(basic_preprocessing)

df.sample(5)

# Toxicity metric distribution problem

(df.ref_tox > df.trn_tox).mean()

We see that the dataset is not prepared yet: the `source` and `reference` fields do not correspond to `toxic` and `non-toxic` accordingly. We will fix this issue in `1.1-data-preparation.ipynb` notebook.
