==> notebooks/1.1-data-preparation.ipynb <==
import pandas as pd

Let's load the data:

df = pd.read_csv('../data/raw/filtered.tsv', delimiter='\t', index_col=0)
df

# Source and Target disambiguation

We see that the fields `reference` and `translation` are often mixed in terms of toxicity level. Let us introduce the fields `source` and `target` and distribute the samples so that `toxicity(source) > toxicity(target)`.

`source` and `target` fields will be a better choice for training and fine-tuning the models. 

columns = ['reference', 'translation', 'ref_tox', 'trn_tox']
df['source'] = df[columns].apply(lambda x: x['reference'] if x['ref_tox'] > x['trn_tox'] else x['translation'], axis=1)
df['target'] = df[columns + ['source']].apply(lambda x: x['translation'] if x['source'] == x['reference'] else x['reference'], axis=1)
df.sample(5)

## Save the data (to /src later)

df.to_csv('../data/interim/distinguished.tsv', sep='\t', index=False)

df_dvizh = pd.read_csv('../data/interim/distinguished.tsv', delimiter='\t')
df_dvizh
