==> notebooks/3.0-t5-finetuning.ipynb <==
In this notebook, we fine-tune and infer a T5 model for paraphrasing using the dataset prepared in `1.1-data-preparation.ipynb` notebook.

import gc
import os
import torch
import pandas as pd

from sklearn.model_selection import train_test_split
from transformers import T5ForConditionalGeneration, T5TokenizerFast, Trainer, TrainingArguments
from transformers.file_utils import cached_property
from torch.utils.data import DataLoader
from typing import List, Dict, Union


os.environ['CUDA_VISIBLE_DEVICES'] = '1'
torch.cuda.is_available()

# Data Preprocessing

Load the dataset and tokenizer:

df = pd.read_csv('../data/interim/distinguished.tsv', delimiter='\t')
df

tokenizer = T5TokenizerFast.from_pretrained("ceshine/t5-paraphrase-paws-msrp-opinosis")

Perform a train-test split:

df_train, df_test_sets = train_test_split(df, test_size=2000)
df_test, df_valid = train_test_split(df_test_sets, test_size=1000)
print(df_train.shape, df_test.shape, df_valid.shape)

Save the validation set to a file. We will use it to evaluate the models in the end of the assignment. 

df_valid.to_csv('../data/interim/validation.tsv', sep='\t')

Apply the tokenization to the data:

x_train = tokenizer(df_train['source'].tolist(), truncation=True)
y_train = tokenizer(df_train['target'].tolist(), truncation=True)
x_test = tokenizer(df_test['source'].tolist(), truncation=True)
y_test = tokenizer(df_test['target'].tolist(), truncation=True)

Prepare a Dataset class and data loaders:

class ParaphraseDataset(torch.utils.data.Dataset):
    def __init__(self, x, y):
        self.x = x
        self.y = y

    def __len__(self):
        return len(self.x['input_ids'])

    def __getitem__(self, idx):
        assert idx < len(self)
        item = {
            key: val[idx] for key, val in self.x.items()
        }
        item['decoder_attention_mask'] = self.y['attention_mask'][idx]
        item['labels'] = self.y['input_ids'][idx]
        return item

train_dataset = ParaphraseDataset(x_train, y_train)
test_dataset = ParaphraseDataset(x_test, y_test)
len(train_dataset), len(test_dataset)

train_dataloader = DataLoader(train_dataset, batch_size=4, drop_last=True, shuffle=True, num_workers=8)
test_dataloader = DataLoader(test_dataset, batch_size=4, drop_last=True, shuffle=True, num_workers=8)

# T5 model fine-tuning for paraphrasing

Let us take the T5 model fine-tuned by Skolkovo Institute for text paraphrasing ([link](https://huggingface.co/s-nlp/t5-paraphrase-paws-msrp-opinosis-paranmt))

- This model is fine-tuned for paraphrasing on PAWS, MSRP, Opinosis, and ParaNMT datasets which is quite a big corpus.
- We will fine-tune it further using the ParaNMT toxicity subset that is given for this assignment (and additionally processed in previous notebooks).

model = T5ForConditionalGeneration.from_pretrained('SkolkovoInstitute/t5-paraphrase-paws-msrp-opinosis-paranmt')

device = torch.device('cuda:0')
model.to(device)

class TrainingArgumentsWrapper(TrainingArguments):
    def __init__(self, **kwargs):
        super().__init__(**kwargs)
        
        self.distributed_state = None

    @cached_property
    def _setup_devices(self):
        return device
    

class DataCollatorWithPadding:
    def __init__(self, tokenizer):
        self.tokenizer = tokenizer

    def __call__(
            self,
            features: List[Dict[str, Union[List[int], torch.Tensor]]]
        ) -> Dict[str, torch.Tensor]:
        
        batch = self.tokenizer.pad(
            features,
            padding=True,
        )
        y_batch = self.tokenizer.pad(
            {
                'input_ids': batch['labels'],
                'attention_mask': batch['decoder_attention_mask']
            },
            padding=True,
        ) 
        batch['labels'] = y_batch['input_ids']
        batch['decoder_attention_mask'] = y_batch['attention_mask']
        
        return {k: torch.tensor(v) for k, v in batch.items()}

dir_name = 't5-paraphrase'

training_args = TrainingArgumentsWrapper(
    output_dir=f'../models/{dir_name}',
    overwrite_output_dir=False,
    num_train_epochs=2,
    per_device_train_batch_size=4,
    gradient_accumulation_steps=4,
    per_device_eval_batch_size=8,
    warmup_steps=300,
    weight_decay=0,
    learning_rate=3e-5,
    logging_dir=f'../models/logs/{dir_name}',
    logging_steps=100,
    eval_steps=100,
    evaluation_strategy='steps',
    save_total_limit=1,
    save_steps=5000,
)

data_collator = DataCollatorWithPadding(tokenizer=tokenizer)

trainer = Trainer(
    model=model,
    args=training_args,
    train_dataset=train_dataset,
    eval_dataset=test_dataset,
    data_collator=data_collator,
    tokenizer=tokenizer,
)

gc.collect()
torch.cuda.empty_cache()

trainer.train()

trainer.evaluate()

Let us infer the resulted model with an example:

model.eval()

toxic_sample = "I'm going to hit you in all directions, civil and criminal, on all counts."

inputs = tokenizer(toxic_sample, return_tensors='pt')
inputs = {
    k: v.to(device) for k, v in inputs.items()
}

for t in model.generate(**inputs, num_return_sequences=10, do_sample=False, num_beams=10):
    print(tokenizer.decode(t, skip_special_tokens=True))

Finally, let's save the model:

model.save_pretrained(f'../models/{dir_name}')
