==> notebooks/4.0-condbert-inference.ipynb <==
In this notebook, we infer the CondBERT model proposed in [the paper](https://arxiv.org/abs/2109.08914).

This model is essentially a zero-shot BERT applied to tokens replacement task. We firstly replace the toxic words with the `[MASK]` token, and then the BERT is inferred to replace the masks with words that fit the meaning.

import sys

sys.path.insert(0, '../')

from src.models.condbert.condbert_inference import CondBERT

condbert = CondBERT(dir_path='../src/models/condbert/')

samples = [
    "What a great sunny day, I miss it.",
    "I didn't fuck him",
    "I'm going to hit you in all directions, civil and criminal, on all counts.",
    "What a fucked rainy day, goddamnit.",
    "hello there! i'm a piece of shit :)",
    "i'm gonna un-fuck your shit uptight, motherfucker!",
]

for sample in samples:
    paraphrase = condbert.detoxify(sentence=sample)
        
    print(f"Sample:\t\t'{sample}'")
    print(f"Paraphrase:\t{paraphrase}" + "\n")

We see this model solves the problem much worse compared to T5, for example.

- Although the toxicity is removed, we often see some artifacts instead of removed words.
- The initial sentence intent may not be preserved, especially in short sentences, where each token carries some semantic weight. In this case, masking of only one token could result in sense missing.
