==> notebooks/5.0-evaluation.ipynb <==
Let us compare the fine-tuned T5 and CondBERT models on the validation set.

Loading the models:

import pandas as pd
import numpy as np
import sys

sys.path.insert(0, '../')

from src.models.condbert.condbert_inference import CondBERT
from src.models.t5_paraphrase.t5_inference import T5Paraphrase
from src.toxicity_metric import measure_toxicity

condbert = CondBERT(dir_path='../src/models/condbert/')

t5 = T5Paraphrase(model_dir_path='../models/t5-paraphrase/')

Reading the data:

df = pd.read_csv('../data/interim/validation.tsv', delimiter='\t', index_col=0)
df.head()

Inference on the validation dataset:

# t5_paraphrase
df['t5'] = df['source'].apply(lambda s: t5.detoxify(sentence=s))

# condbert
df['condbert'] = df['source'].apply(lambda s: condbert.detoxify(sentence=s))

df = pd.read_csv('../data/interim/predictions.tsv', delimiter='\t', index_col=0)
df.head()

Now, let's measure the average toxicity level, and toxicity percentage for the `source`, `target`, `t5`, and `condbert` columns:

results_dict = {
    'toxicity_percentage': {},
    'avg_toxicity': {}
}

for column in ['source', 'target', 't5', 'condbert']:
    toxicity_labels, toxicity_scores = measure_toxicity(df[column].to_list())
    
    results_dict['toxicity_percentage'][column] = np.sum(toxicity_labels) / len(toxicity_labels)
    results_dict['avg_toxicity'][column] = np.mean(toxicity_scores)

df_results = pd.DataFrame.from_dict(results_dict)
df_results

As you can see, the condbert model is the best one according to the introduced metrics :)

However, we have no fluency metric which is also an important one when talking about text seq2seq task. Condbert's generations often look clumsy and distort the initial message meaning.
So, if I have managed to implement some fluency metric, the condbert would certainly perform poorly with it.

We see that t5 generations on average has less toxicity compared to the `target` column (6.22 > 5.57). However, t5's toxicity percentage is slightly greater than the target's one.
